edited test cases
